requirements.txt
# Project dependencies
# requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspread
pandas
python-dotenv
requests
tweepy
openai
genai
google-generativeai
# Note: Facebook and LinkedIn often use direct HTTP requests with 'requests'